Move remaining theme-related code and tests from settings mod to theme mod
